chore: move file lock tests into test_unit.py
